requirement.txt
pandas
matplotlib
plotly-express
seaborn
scikit-learn
yfinance
pandas_ta
pandas_datareader
scipy
numpy
ydata-profiling
statsmodels
scikeras
kaleido
tensorflow
keras-tuner
pdfkit
ipywidgets
IPython
